split main.dart into bootstrap and app entry files
